reworked entire debugging window. hold shift while scrolling to scroll pages instead of lines (mouse needs to hover instructions entries). click on entry to set breakpoint, cpu runs until it reaches the breakpoint and then stops executing

+++ gameboyx/GraphicsUnitSM83.cpp
@@ -107,7 +107,7 @@ void GraphicsUnitSM83::DrawTileBackground(const int& _pos_x, const int& _pos_y) 
 // draw window (window pos at top left corner is (WX-7/WY) !)
 void GraphicsUnitSM83::DrawTileMapWindow() {
 	u8 win_offset = graphics_ctx->win_tilemap_offset;
-	int wx = graphics_ctx->WX - 7;
+	int wx = (int)graphics_ctx->WX - 7;
 	int wy = graphics_ctx->WY;
 
 	for (int x = 0; x < PPU_TILES_HORIZONTAL - wx; x+=PPU_PIXELS_TILE_X) {
